konekt/requirements.txt
alabaster==0.7.12
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
Babel==2.9.1
beautifulsoup4==4.9.3
blis==0.7.4
bootstrap-py==1.0.2
boto3==1.17.55
botocore==1.20.55
bs4==0.0.1
bump2version==1.0.1
bumpversion==0.6.0
catalogue==2.0.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
cymem==2.0.5
defusedxml==0.7.1
Django==3.1.6
django-elasticsearch-dsl==7.1.4
django-elasticsearch-dsl-drf==0.21
django-nine==0.2.4
djangorestframework==3.12.2
docutils==0.16
docx-mailmerge==0.5.0
elasticsearch==7.10.1
elasticsearch-dsl==7.3.0
environs==9.3.2
Faker==8.1.0
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.0.0/fr_core_news_sm-3.0.0-py3-none-any.whl
gitdb==4.0.7
GitPython==3.1.14
greenlet==1.0.0
idna==2.10
imagesize==1.2.0
importlib-metadata==4.0.1
iniconfig==1.1.1
Jinja2==2.11.3
jmespath==0.10.0
lxml==4.6.2
MarkupSafe==1.1.1
marshmallow==3.11.1
murmurhash==1.0.5
mysqlclient==2.0.3
newrelic==6.2.0.156
numpy==1.20.1
packaging==20.9
pandas==1.2.1
pathy==0.5.2
pdf2image==1.15.1
pguard==0.4.0
Pillow==8.2.0
pluggy==0.13.1
preshed==3.0.5
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.10.0
pydantic==1.7.3
Pygments==2.9.0
pyparsing==2.4.7
pytesseract==0.3.7
pytest==6.2.2
pytest-django==4.1.0
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-aws4auth==1.0.1
s3transfer==0.4.1
selenium==3.141.0
six==1.15.0
smart-open==3.0.0
smmap==4.0.0
snowballstemmer==2.1.0
soupsieve==2.1
spacy==3.0.5
spacy-legacy==3.0.5
Sphinx==3.5.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.4.10
sqlparse==0.4.1
srsly==2.4.1
tesseract==0.1.3
text-unidecode==1.3
thinc==8.0.3
tika==1.24
toml==0.10.2
tqdm==4.60.0
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wasabi==0.8.2
webdriver-manager==3.3.0
zipp==3.4.1
